Exception handling - delete

+++ src/main/java/com/springbootproject/demo/services/UserService.java
@@ -2,8 +2,11 @@ package com.springbootproject.demo.services;
 
 import com.springbootproject.demo.entities.User;
 import com.springbootproject.demo.repositories.UserRepository;
+import com.springbootproject.demo.services.exceptions.DatabaseException;
 import com.springbootproject.demo.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +31,14 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+
+        try {
+            if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
